--- src/app/api/my/reservation/route.ts
@@ -1,62 +1,0 @@
-import axiosInstance from "@/lib/api/axiosInstanceApi";
-import { isAxiosError } from "axios";
-import { NextRequest, NextResponse } from "next/server";
-
-// 내 예약 요청
-export const GET = async (req: NextRequest) => {
-  const accessToken = req.cookies.get("accessToken")?.value;
-
-  if (!accessToken) {
-    return NextResponse.json({ message: "로그인이 필요합니다" }, { status: 401 });
-  }
-
-  try {
-    // const url = new URL(req.url);
-    const cursorId = req.nextUrl.searchParams.get("cursorId");
-    const status = req.nextUrl.searchParams.get("status");
-    console.log("status", status);
-
-    let url = `/my-reservations?size=10&status=${status}`;
-
-    if (cursorId) url += `&cursorId=${cursorId}`;
-    const response = await axiosInstance.get(url, {
-      headers: {
-        Authorization: `Bearer ${accessToken}`,
-      },
-    });
-    return NextResponse.json(response.data, { status: response.status });
-  } catch (error: unknown) {
-    if (isAxiosError(error)) {
-      const errorMessage = error.response?.data?.message || "서버에서 알 수 없는 오류가 발생했습니다.";
-      return NextResponse.json({ message: errorMessage }, { status: error.response?.status || 500 });
-    }
-
-    return NextResponse.json({ message: "서버에서 알 수 없는 오류가 발생했습니다." }, { status: 500 });
-  }
-};
-
-// 내 예약 취소 요청
-export const PATCH = async (req: NextRequest) => {
-  const accessToken = req.cookies.get("accessToken")?.value;
-
-  if (!accessToken) {
-    return NextResponse.json({ message: "로그인이 필요합니다" }, { status: 401 });
-  }
-
-  try {
-    const body = await req.json();
-    const response = await axiosInstance.patch(`/my-reservations/${body.reservationId}`, body, {
-      headers: {
-        Authorization: `Bearer ${accessToken}`,
-      },
-    });
-    return NextResponse.json(response.data, { status: response.status });
-  } catch (error: unknown) {
-    if (isAxiosError(error)) {
-      const errorMessage = error.response?.data?.message || "서버에서 알 수 없는 오류가 발생했습니다.";
-      return NextResponse.json({ message: errorMessage }, { status: error.response?.status || 500 });
-    }
-
-    return NextResponse.json({ message: "서버에서 알 수 없는 오류가 발생했습니다." }, { status: 500 });
-  }
-};
